refactor: streamline column update logic and emit event in ColumnSettings
